Disable solana and enable eclipse sol wallets
