+++ package.json
@@ -26,7 +26,7 @@
     "@project-serum/anchor": "^0.26.0",
     "@reservoir0x/relay-bitcoin-wallet-adapter": "^1.0.6",
     "@reservoir0x/relay-kit-hooks": "1.4.7",
-    "@reservoir0x/relay-kit-ui": "2.2.15",
+    "@reservoir0x/relay-kit-ui": "2.2.18",
     "@reservoir0x/relay-sdk": "1.4.6",
     "@reservoir0x/relay-svm-wallet-adapter": "^2.0.6",
     "@solana/spl-token": "^0.4.9",
